refactor: transforming the generic category query search into a useful class for pagination
