+++ catalog-admin/domain/src/main/java/com/github/lucasbarbosaalves/catalog/domain/category/CategoryGateway.java
@@ -1,5 +1,6 @@
 package com.github.lucasbarbosaalves.catalog.domain.category;
 
+import com.github.lucasbarbosaalves.catalog.domain.pagination.SearchQuery;
 import com.github.lucasbarbosaalves.catalog.domain.pagination.Pagination;
 
 import java.util.Optional;
@@ -14,6 +15,6 @@ public interface CategoryGateway {
 
     Category update(Category category);
 
-    Pagination<Category> findAll(CategorySearchQuery query);
+    Pagination<Category> findAll(SearchQuery query);
 
 }
